+++ src/controllers/system/database.ts
@@ -1,8 +1,7 @@
 import {requireAuth, requireValidUser, requireAdmin} from '../middlewares/auth';
 import {generateDB} from '../../dao/database';
 import { Router } from 'express';
-import { dumpPG } from '../../utils/postgresql';
-import { resetViewcounts } from '../../dao/kara';
+import { dumpPG, restorePG } from '../../utils/postgresql';
 import { requireNotDemo } from '../middlewares/demo';
 import { updateMedias } from '../../services/download';
 import { getConfig } from '../../lib/utils/config';
@@ -26,19 +25,17 @@ export default function systemDBController(router: Router) {
 	router.post('/system/db/dump', requireAuth, requireValidUser, requireAdmin, async (_req: any, res: any) => {
 		try {
 			await dumpPG();
-			res.status(200).send('Database dumped to karaokemugen.pgdump');
+			res.status(200).send('Database dumped to karaokemugen.sql');
 		} catch(err) {
 			res.status(500).send(`Error dumping database : ${err}`);
 		}
 	});
-
-	router.post('/system/db/resetviewcounts', requireAuth, requireValidUser, requireAdmin, async (_req: any, res: any) => {
+	router.post('/system/db/restore', requireAuth, requireValidUser, requireAdmin, async (_req: any, res: any) => {
 		try {
-			await resetViewcounts();
-			res.status(200).send('Viewcounts successfully reset');
+			await restorePG();
+			res.status(200).send('Database restored from karaokemugen.sql');
 		} catch(err) {
-			res.status(500).send(`Error resetting viewcounts: ${err}`)
+			res.status(500).send(`Error restoring database : ${err}`);
 		}
 	});
-
 }