+++ src/types/binChecker.d.ts
@@ -3,5 +3,6 @@ export interface BinariesConfig {
 	mpv: string,
 	postgres: string,
 	postgres_ctl: string,
-	postgres_dump: string
+	postgres_dump: string,
+	postgres_client: string
 }